Split snippets into different files.
