+++ compose/snippets/src/main/java/com/example/compose/snippets/animations/sharedelement/BasicSharedElementSnippets.kt
@@ -1,0 +1,467 @@
+@file:OptIn(ExperimentalSharedTransitionApi::class)
+
+package com.example.compose.snippets.animations.sharedelement
+
+import androidx.compose.animation.AnimatedContent
+import androidx.compose.animation.AnimatedVisibilityScope
+import androidx.compose.animation.ExperimentalSharedTransitionApi
+import androidx.compose.animation.SharedTransitionLayout
+import androidx.compose.animation.SharedTransitionScope
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.alpha
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.compose.snippets.R
+import com.example.compose.snippets.ui.theme.LavenderLight
+import com.example.compose.snippets.ui.theme.RoseLight
+
+private class SharedElementBasicUsage1 {
+    @Preview
+    // [START android_compose_animations_shared_element_start]
+    @Composable
+    private fun SharedElementApp() {
+        var showDetails by remember {
+            mutableStateOf(false)
+        }
+        AnimatedContent(
+            showDetails,
+            label = "basic_transition"
+        ) { targetState ->
+            if (!targetState) {
+                MainContent(onShowDetails = {
+                    showDetails = true
+                })
+            } else {
+                DetailsContent(onBack = {
+                    showDetails = false
+                })
+            }
+        }
+    }
+
+    @Composable
+    private fun MainContent(onShowDetails: () -> Unit) {
+        Row(
+            // [START_EXCLUDE]
+            modifier = Modifier
+                .padding(8.dp)
+                .border(1.dp, Color.Gray.copy(alpha = 0.5f), RoundedCornerShape(8.dp))
+                .background(LavenderLight, RoundedCornerShape(8.dp))
+                .clickable {
+                    onShowDetails()
+                }
+                .padding(8.dp)
+            // [END_EXCLUDE]
+        ) {
+            Image(
+                painter = painterResource(id = R.drawable.cupcake),
+                contentDescription = "Cupcake",
+                modifier = Modifier
+                    .size(100.dp)
+                    .clip(CircleShape),
+                contentScale = ContentScale.Crop
+            )
+            Text("Cupcake", fontSize = 21.sp)
+        }
+    }
+
+    @Composable
+    private fun DetailsContent(modifier: Modifier = Modifier, onBack: () -> Unit) {
+        Column(
+            // [START_EXCLUDE]
+            modifier = Modifier
+                .padding(top = 200.dp, start = 16.dp, end = 16.dp)
+                .border(1.dp, Color.Gray.copy(alpha = 0.5f), RoundedCornerShape(8.dp))
+                .background(RoseLight, RoundedCornerShape(8.dp))
+                .clickable {
+                    onBack()
+                }
+                .padding(8.dp)
+            // [END_EXCLUDE]
+        ) {
+            Image(
+                painter = painterResource(id = R.drawable.cupcake),
+                contentDescription = "Cupcake",
+                modifier = Modifier
+                    .size(200.dp)
+                    .clip(CircleShape),
+                contentScale = ContentScale.Crop
+            )
+            Text("Cupcake", fontSize = 28.sp)
+            // [START_EXCLUDE]
+            Text(
+                "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur sit amet lobortis velit. " +
+                        "Lorem ipsum dolor sit amet, consectetur adipiscing elit." +
+                        " Curabitur sagittis, lectus posuere imperdiet facilisis, nibh massa " +
+                        "molestie est, quis dapibus orci ligula non magna. Pellentesque rhoncus " +
+                        "hendrerit massa quis ultricies. Curabitur congue ullamcorper leo, at maximus"
+            )
+            // [END_EXCLUDE]
+        }
+    }
+    // [END android_compose_animations_shared_element_start]
+}
+
+private class SharedElementBasicUsage2 {
+    @Preview
+    @Composable
+    private fun SharedElementApp() {
+        // [START android_compose_animations_shared_element_step1]
+        var showDetails by remember {
+            mutableStateOf(false)
+        }
+        SharedTransitionLayout {
+            AnimatedContent(
+                showDetails,
+                label = "basic_transition"
+            ) { targetState ->
+                if (!targetState) {
+                    MainContent(
+                        onShowDetails = {
+                            showDetails = true
+                        },
+                        animatedVisibilityScope = this@AnimatedContent,
+                        sharedTransitionScope = this@SharedTransitionLayout
+                    )
+                } else {
+                    DetailsContent(
+                        onBack = {
+                            showDetails = false
+                        },
+                        animatedVisibilityScope = this@AnimatedContent,
+                        sharedTransitionScope = this@SharedTransitionLayout
+                    )
+                }
+            }
+        }
+        // [END android_compose_animations_shared_element_step1]
+    }
+
+    @Composable
+    private fun MainContent(
+        onShowDetails: () -> Unit,
+        sharedTransitionScope: SharedTransitionScope,
+        animatedVisibilityScope: AnimatedVisibilityScope
+    ) {
+        Row(
+            // [START_EXCLUDE]
+            modifier = Modifier
+                .padding(8.dp)
+                .border(1.dp, Color.Gray.copy(alpha = 0.5f), RoundedCornerShape(8.dp))
+                .background(LavenderLight, RoundedCornerShape(8.dp))
+                .clickable {
+                    onShowDetails()
+                }
+                .padding(8.dp)
+            // [END_EXCLUDE]
+        ) {
+            Image(
+                painter = painterResource(id = R.drawable.cupcake),
+                contentDescription = "Cupcake",
+                modifier = Modifier
+                    .size(100.dp)
+                    .clip(CircleShape),
+                contentScale = ContentScale.Crop
+            )
+            Text("Cupcake", fontSize = 21.sp)
+        }
+    }
+
+    @Composable
+    private fun DetailsContent(
+        modifier: Modifier = Modifier,
+        onBack: () -> Unit,
+        sharedTransitionScope: SharedTransitionScope,
+        animatedVisibilityScope: AnimatedVisibilityScope
+    ) {
+        Column(
+            // [START_EXCLUDE]
+            modifier = Modifier
+                .padding(top = 200.dp, start = 16.dp, end = 16.dp)
+                .border(1.dp, Color.Gray.copy(alpha = 0.5f), RoundedCornerShape(8.dp))
+                .background(RoseLight, RoundedCornerShape(8.dp))
+                .clickable {
+                    onBack()
+                }
+                .padding(8.dp)
+            // [END_EXCLUDE]
+        ) {
+            Image(
+                painter = painterResource(id = R.drawable.cupcake),
+                contentDescription = "Cupcake",
+                modifier = Modifier
+                    .size(200.dp)
+                    .clip(CircleShape),
+                contentScale = ContentScale.Crop
+            )
+            Text("Cupcake", fontSize = 28.sp)
+            // [START_EXCLUDE]
+            Text(
+                "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur sit amet lobortis velit. " +
+                        "Lorem ipsum dolor sit amet, consectetur adipiscing elit." +
+                        " Curabitur sagittis, lectus posuere imperdiet facilisis, nibh massa " +
+                        "molestie est, quis dapibus orci ligula non magna. Pellentesque rhoncus " +
+                        "hendrerit massa quis ultricies. Curabitur congue ullamcorper leo, at maximus"
+            )
+            // [END_EXCLUDE]
+        }
+    }
+}
+
+
+
+private class SharedElementBasicUsage3 {
+
+    @Preview
+    @Composable
+    private fun SharedElementApp() {
+        var showDetails by remember {
+            mutableStateOf(false)
+        }
+        SharedTransitionLayout {
+            AnimatedContent(
+                showDetails,
+                label = "basic_transition"
+            ) { targetState ->
+                if (!targetState) {
+                    MainContent(
+                        onShowDetails = {
+                            showDetails = true
+                        },
+                        animatedVisibilityScope = this@AnimatedContent,
+                        sharedTransitionScope = this@SharedTransitionLayout
+                    )
+                } else {
+                    DetailsContent(
+                        onBack = {
+                            showDetails = false
+                        },
+                        animatedVisibilityScope = this@AnimatedContent,
+                        sharedTransitionScope = this@SharedTransitionLayout
+                    )
+                }
+            }
+        }
+    }
+
+    // [START android_compose_animations_shared_element_step2]
+    @Composable
+    private fun MainContent(
+        onShowDetails: () -> Unit,
+        modifier: Modifier = Modifier,
+        sharedTransitionScope: SharedTransitionScope,
+        animatedVisibilityScope: AnimatedVisibilityScope
+    ) {
+        Row(
+            // [START_EXCLUDE]
+            modifier = Modifier
+                .padding(8.dp)
+                .border(1.dp, Color.Gray.copy(alpha = 0.5f), RoundedCornerShape(8.dp))
+                .background(LavenderLight, RoundedCornerShape(8.dp))
+                .clickable {
+                    onShowDetails()
+                }
+                .padding(8.dp)
+            // [END_EXCLUDE]
+        ) {
+            with(sharedTransitionScope) {
+                Image(
+                    painter = painterResource(id = R.drawable.cupcake),
+                    contentDescription = "Cupcake",
+                    modifier = Modifier
+                        .sharedElement(
+                            rememberSharedContentState(key = "image"),
+                            animatedVisibilityScope = animatedVisibilityScope
+                        )
+                        .size(100.dp)
+                        .clip(CircleShape),
+                    contentScale = ContentScale.Crop
+                )
+                Text(
+                    "Cupcake", fontSize = 21.sp,
+                    modifier = Modifier.sharedBounds(
+                        rememberSharedContentState(key = "title"),
+                        animatedVisibilityScope = animatedVisibilityScope
+                    )
+                )
+            }
+        }
+    }
+
+    @Composable
+    private fun DetailsContent(
+        modifier: Modifier = Modifier,
+        onBack: () -> Unit,
+        sharedTransitionScope: SharedTransitionScope,
+        animatedVisibilityScope: AnimatedVisibilityScope
+    ) {
+        Column(
+            // [START_EXCLUDE]
+            modifier = Modifier
+                .padding(top = 200.dp, start = 16.dp, end = 16.dp)
+                .border(1.dp, Color.Gray.copy(alpha = 0.5f), RoundedCornerShape(8.dp))
+                .background(RoseLight, RoundedCornerShape(8.dp))
+                .clickable {
+                    onBack()
+                }
+                .padding(8.dp)
+            // [END_EXCLUDE]
+        ) {
+            with(sharedTransitionScope) {
+                Image(
+                    painter = painterResource(id = R.drawable.cupcake),
+                    contentDescription = "Cupcake",
+                    modifier = Modifier
+                        .sharedElement(
+                            rememberSharedContentState(key = "image"),
+                            animatedVisibilityScope = animatedVisibilityScope
+                        )
+                        .size(200.dp)
+                        .clip(CircleShape),
+                    contentScale = ContentScale.Crop
+                )
+                Text(
+                    "Cupcake", fontSize = 28.sp,
+                    modifier = Modifier.sharedBounds(
+                        rememberSharedContentState(key = "title"),
+                        animatedVisibilityScope = animatedVisibilityScope
+                    )
+                )
+                // [START_EXCLUDE]
+                Text(
+                    "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur sit amet lobortis velit. " +
+                            "Lorem ipsum dolor sit amet, consectetur adipiscing elit." +
+                            " Curabitur sagittis, lectus posuere imperdiet facilisis, nibh massa " +
+                            "molestie est, quis dapibus orci ligula non magna. Pellentesque rhoncus " +
+                            "hendrerit massa quis ultricies. Curabitur congue ullamcorper leo, at maximus"
+                )
+                // [END_EXCLUDE]
+            }
+        }
+    }
+    // [END android_compose_animations_shared_element_step2]
+}
+
+
+@Preview
+@Composable
+private fun SharedElement_ManualVisibleControl() {
+    // [START android_compose_shared_element_manual_control]
+    var selectFirst by remember { mutableStateOf(true) }
+    val key = remember { Any() }
+    SharedTransitionLayout(
+        Modifier
+            .fillMaxSize()
+            .padding(10.dp)
+            .clickable {
+                selectFirst = !selectFirst
+            }
+    ) {
+        Box(
+            Modifier
+                .sharedElementWithCallerManagedVisibility(
+                    rememberSharedContentState(key = key),
+                    !selectFirst
+                )
+                .background(Color.Red)
+                .size(100.dp)
+        ) {
+            Text(if (!selectFirst) "false" else "true", color = Color.White)
+        }
+        Box(
+            Modifier
+                .offset(180.dp, 180.dp)
+                .sharedElementWithCallerManagedVisibility(
+                    rememberSharedContentState(
+                        key = key,
+                    ),
+                    selectFirst
+                )
+                .alpha(0.5f)
+                .background(Color.Blue)
+                .size(180.dp)
+        ) {
+            Text(if (selectFirst) "false" else "true", color = Color.White)
+        }
+    }
+    // [END android_compose_shared_element_manual_control]
+}
+
+@Preview
+@Composable
+private fun UnmatchedBoundsExample() {
+    // [START android_compose_animation_shared_element_bounds_unmatched]
+    var selectFirst by remember { mutableStateOf(true) }
+    val key = remember { Any() }
+    SharedTransitionLayout(
+        Modifier
+            .fillMaxSize()
+            .padding(10.dp)
+            .clickable {
+                selectFirst = !selectFirst
+            }
+    ) {
+        AnimatedContent(targetState = selectFirst, label = "AnimatedContent") { targetState ->
+            if (targetState) {
+                Box(
+                    Modifier
+                        .padding(12.dp)
+                        .sharedBounds(
+                            rememberSharedContentState(key = key),
+                            animatedVisibilityScope = this@AnimatedContent
+                        )
+                        .border(2.dp, Color.Red)
+                ) {
+                    Text(
+                        "Hello",
+                        fontSize = 20.sp
+                    )
+                }
+            } else {
+                Box(
+                    Modifier
+                        .offset(180.dp, 180.dp)
+                        .sharedBounds(
+                            rememberSharedContentState(
+                                key = key,
+                            ),
+                            animatedVisibilityScope = this@AnimatedContent
+                        )
+                        .border(2.dp, Color.Red)
+                        .padding(12.dp)
+
+                ) {
+                    Text(
+                        "Hello",
+                        fontSize = 36.sp
+                    )
+                }
+            }
+        }
+    }
+    // [END android_compose_animation_shared_element_bounds_unmatched]
+}
