+++ compose/snippets/src/main/java/com/example/compose/snippets/animations/sharedelement/SharedElementCommonUseCaseSnippets.kt
@@ -1,0 +1,84 @@
+@file:OptIn(ExperimentalSharedTransitionApi::class)
+
+package com.example.compose.snippets.animations.sharedelement
+
+import androidx.annotation.DrawableRes
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.ExitTransition
+import androidx.compose.animation.ExperimentalSharedTransitionApi
+import androidx.compose.animation.SharedTransitionLayout
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.wrapContentWidth
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalInspectionMode
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImage
+import coil.request.ImageRequest
+
+
+@Preview
+@Composable
+private fun SharedAsyncImage() {
+    SharedTransitionLayout {
+        AnimatedVisibility(visible = true) {
+            // [START android_compose_shared_element_async_image_tip]
+            AsyncImage(
+                model = ImageRequest.Builder(LocalContext.current)
+                    .data("your-image-url")
+                    .crossfade(true)
+                    .build(),
+                placeholder = null,
+                contentDescription = null,
+                modifier = Modifier
+                    .size(120.dp)
+                    .sharedBounds(
+                        rememberSharedContentState(
+                            key = "image-key"
+                        ),
+                        animatedVisibilityScope = this,
+                        exit = ExitTransition.None
+                    )
+            )
+            // [END android_compose_shared_element_async_image_tip]
+        }
+    }
+}
+
+@Composable
+fun debugPlaceholder(@DrawableRes debugPreview: Int) =
+    if (LocalInspectionMode.current) {
+        painterResource(id = debugPreview)
+    } else {
+        null
+    }
+
+@Preview
+@Composable
+private fun SharedElementTypicalUseText() {
+    SharedTransitionLayout {
+        AnimatedVisibility(visible = true) {
+            // [START android_compose_shared_element_text_tip]
+            Text(
+                text = "This is an example of how to share text",
+                modifier = Modifier
+                    .wrapContentWidth()
+                    .sharedBounds(
+                        rememberSharedContentState(
+                            key = "shared Text"
+                        ),
+                        animatedVisibilityScope = this,
+                        enter = fadeIn() + scaleInSharedContentToBounds(),
+                        exit = fadeOut() + scaleOutSharedContentToBounds()
+                    )
+            )
+            // [END android_compose_shared_element_text_tip]
+        }
+    }
+}
