+++ compose/snippets/src/main/java/com/example/compose/snippets/animations/sharedelement/SharedElementsWithNavigationSnippets.kt
@@ -1,0 +1,152 @@
+@file:OptIn(ExperimentalSharedTransitionApi::class)
+
+package com.example.compose.snippets.animations.sharedelement
+
+import androidx.annotation.DrawableRes
+import androidx.compose.animation.ExperimentalSharedTransitionApi
+import androidx.compose.animation.SharedTransitionLayout
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.aspectRatio
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.itemsIndexed
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.geometry.Rect
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.navigation.NavType
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import androidx.navigation.navArgument
+import com.example.compose.snippets.R
+
+
+val listSnacks = listOf(
+    Snack("Cupcake", "", R.drawable.cupcake),
+    Snack("Donut", "", R.drawable.donut),
+    Snack("Eclair", "", R.drawable.eclair),
+    Snack("Froyo", "", R.drawable.froyo),
+    Snack("Gingerbread", "", R.drawable.gingerbread),
+    Snack("Honeycomb", "", R.drawable.honeycomb),
+)
+
+@Preview
+@Composable
+fun SharedElement_PredictiveBack() {
+    // [START android_compose_shared_element_predictive_back]
+    SharedTransitionLayout {
+        val boundsTransform = { _: Rect, _: Rect -> tween<Rect>(1400) }
+
+        val navController = rememberNavController()
+        NavHost(
+            navController = navController,
+            startDestination = "home"
+        ) {
+
+            composable("home") {
+                LazyColumn(
+                    modifier = Modifier
+                        .fillMaxSize()
+                        .padding(8.dp),
+                    verticalArrangement = Arrangement.spacedBy(8.dp)
+                ) {
+                    itemsIndexed(listSnacks) { index, item ->
+                        Row(
+                            Modifier.clickable {
+                                navController.navigate("details/$index")
+                            }
+                        ) {
+                            Spacer(modifier = Modifier.width(8.dp))
+                            Image(
+                                painterResource(id = item.image),
+                                contentDescription = item.description,
+                                contentScale = ContentScale.Crop,
+                                modifier = Modifier
+                                    .sharedElement(
+                                        rememberSharedContentState(key = "image-$index"),
+                                        animatedVisibilityScope = this@composable,
+                                        boundsTransform = boundsTransform
+                                    )
+                                    .size(100.dp)
+                            )
+                            Spacer(modifier = Modifier.width(8.dp))
+                            Text(
+                                item.name, fontSize = 18.sp,
+                                modifier = Modifier
+                                    .align(Alignment.CenterVertically)
+                                    .sharedElement(
+                                        rememberSharedContentState(key = "text-$index"),
+                                        animatedVisibilityScope = this@composable,
+                                        boundsTransform = boundsTransform
+                                    )
+                            )
+                        }
+                    }
+                }
+            }
+            composable(
+                "details/{item}",
+                arguments = listOf(navArgument("item") { type = NavType.IntType })
+            ) { backStackEntry ->
+                val id = backStackEntry.arguments?.getInt("item")
+                val snack = listSnacks[id!!]
+                Column(
+                    Modifier
+                        .fillMaxSize()
+                        .clickable {
+                            navController.navigate("home")
+                        }
+                ) {
+                    Image(
+                        painterResource(id = snack.image),
+                        contentDescription = snack.description,
+                        contentScale = ContentScale.Crop,
+                        modifier = Modifier
+                            .sharedElement(
+                                rememberSharedContentState(key = "image-$id"),
+                                animatedVisibilityScope = this@composable,
+                                boundsTransform = boundsTransform
+                            )
+                            .aspectRatio(1f)
+                            .fillMaxWidth()
+                    )
+                    Text(
+                        snack.name, fontSize = 18.sp,
+                        modifier =
+                        Modifier
+                            .sharedElement(
+                                rememberSharedContentState(key = "text-$id"),
+                                animatedVisibilityScope = this@composable,
+                                boundsTransform = boundsTransform
+                            )
+                            .fillMaxWidth()
+                    )
+                }
+            }
+        }
+    }
+// [END android_compose_shared_element_predictive_back]
+}
+
+data class Snack(
+    val name: String,
+    val description: String,
+    @DrawableRes val image: Int
+)
